36. 식당 리뷰리스트 불러오기

+++ app/src/main/java/com/yb/part6_chapter01/model/restaurant/review/RestaurantReviewModel.kt
@@ -10,6 +10,6 @@ data class RestaurantReviewModel(
     val restaurantTitle: String,
     val title: String,
     val content: String,
-    val grade: Int,
+    val grade: Float,
     val thumbnailImageUri: String? = null,
 ) : Model(id, type)
